Code First Pass
**Description**
This update introduces code to scrape the thousand trails website and
build a CSV of all the resorts available for reservation.

This is an educational experience in project layout and python usage.
This first pass is a little rough and will be cleaned up at a later
date. Tests will be added at a later update once the data model and
module code is cleaned up.