Adding tests and working on project layout
**Description**
When running pytest the main function failed to import local
dependencies. I had to set them as relative imports in order to allow
the tests to execute. This required that I add a app.py in the root
directory to function as the main execution method. Therefore I move
script to bin to be more in line with what I understand the project
structure should look like for a python project.

In addition I broke up some of the code into smaller modules to permit
better testing. It is a bit overkill for the purpose of the script but
it shows a better project layout.