adds campendium rating and cellmaper link (#4)
# Description
This update introduces a helper for generating a link to directly to the att and Verizon cell mapper maps. To support this I leveraged the Google Maps geolocate API to convert the campground address to GPS coordinates that can then be used to build the cell mapper link. 

This also adds the ability to scrape the compendium website for user-reported cell reception.